Traktor: Set rigid bodies enabled after owner entity has been attached to component, thus able to properly place body in world. This was causing a major stall when creating scene.
[git-p4: depot-paths = "//depot/traktor/": change = 21480]

+++ code/Physics/Bullet/PhysicsManagerBullet.cpp
@@ -1278,7 +1278,7 @@ void PhysicsManagerBullet::solveConstraints(const RefArray< Body >& bodies, cons
 		0,
 		0,
 		btConstraints,
-		joints.size(),
+		(int)joints.size(),
 		m_dynamicsWorld->getSolverInfo(),
 		0,
 		m_dynamicsWorld->getDispatcher()
